Move utils and add trace serialization
